works(hopefuIIy)

+++ strategoAI/src/abstractSearchComponents/Playthrough.java
@@ -18,7 +18,7 @@ public class Playthrough<State extends SearchState, Action extends AbstractActio
 	
 	
 
-	public int returnPlaythroughResult(TreeNode<State, Action> leafNode) {
+	public double returnPlaythroughResult(TreeNode<State, Action> leafNode) {
 		@SuppressWarnings("unchecked")
 
 		State state = (State) leafNode.getState().deepCopySelf();
