+++ strategoAI/src/abstractSearchComponents/Selection.java
@@ -26,7 +26,7 @@ public class Selection<State extends SearchState, Action extends AbstractAction<
 		// uct attempt
 		TreeNode<State, Action> tempNode = null;
 		double selectionReferance = 0;
-		int tottalWins = 1;
+		double tottalWins = 1;
 		// System.out.println(selectionReferance);
 		// System.out.println(tottalWins);
 		for (int k = 0; k < childrenList.size(); k++) {
